edit float

+++ core/app/models/view/report/LossRate.java
@@ -53,7 +53,7 @@ public class LossRate implements Serializable {
     /**
      * 总赔偿金额
      */
-    public Double totalamt;
+    public Float totalamt;
 
     /**
      * 货币
@@ -91,7 +91,7 @@ public class LossRate implements Serializable {
         this.totalqty = new BigDecimal(0);
         this.shipqty = new BigDecimal(0);
         this.lossrate = new BigDecimal(0.0);
-        this.totalamt = 0d;
+        this.totalamt = 0f;
         this.payrate = new BigDecimal(0.0);
     }
